+++ src/components/interactables/form/FormControl.svelte
@@ -1,41 +1,32 @@
 <script lang="ts">
     // TODO: Stories (?)
 
-    import {CONTEXT_FORM_ID, make_id_context} from "../../../lib/stores/id";
+    import type {IGlobalProperties} from "../../../lib/types/global";
+    import type {IHTML5Properties} from "../../../lib/types/html5";
+    import type {IMarginProperties} from "../../../lib/types/spacings";
 
-    import type {DESIGN_HIDDEN_ARGUMENT} from "../../../lib/types/hidden";
-    import type {DESIGN_SPACING_ARGUMENT} from "../../../lib/types/spacings";
+    import {CONTEXT_FORM_ID, make_id_context} from "../../../lib/stores/id";
 
     import {map_global_attributes} from "../../../lib/util/attributes";
 
-    export let element: HTMLElement | null = null;
+    type $$Props = {
+        element?: HTMLDivElement;
 
-    let _class: string = "";
-    export let id: string = "";
-    export let name: string = "";
-    export let style: string = "";
-    export let tabindex: number | string = "";
-    export let title: string = "";
+        logic_id?: string;
+    } & IHTML5Properties &
+        IGlobalProperties &
+        IMarginProperties;
 
+    export let element: $$Props["element"] = undefined;
+
+    let _class: $$Props["class"] = "";
     export {_class as class};
 
-    export let hidden: DESIGN_HIDDEN_ARGUMENT = false;
+    export let logic_id: $$Props["logic_id"] = "";
 
-    export let margin: DESIGN_SPACING_ARGUMENT | undefined = undefined;
+    const _logic_id = make_id_context(logic_id as string, CONTEXT_FORM_ID);
 
-    export let margin_x: DESIGN_SPACING_ARGUMENT | undefined = undefined;
-    export let margin_y: DESIGN_SPACING_ARGUMENT | undefined = undefined;
-
-    export let margin_top: DESIGN_SPACING_ARGUMENT | undefined = undefined;
-    export let margin_left: DESIGN_SPACING_ARGUMENT | undefined = undefined;
-    export let margin_bottom: DESIGN_SPACING_ARGUMENT | undefined = undefined;
-    export let margin_right: DESIGN_SPACING_ARGUMENT | undefined = undefined;
-
-    export let logic_id: string = "";
-
-    const _logic_id = make_id_context(logic_id, CONTEXT_FORM_ID);
-
-    $: $_logic_id = logic_id;
+    $: $_logic_id = logic_id as string;
 </script>
 
 <div bind:this={element} {...map_global_attributes($$props)} class="form-control {_class}">