+++ src/components/layouts/grid/GridItem.svelte
@@ -1,25 +1,27 @@
 <script lang="ts">
-    import type {DESIGN_HIDDEN_ARGUMENT} from "../../../lib/types/hidden";
+    import type {IGlobalProperties} from "../../../lib/types/global";
+    import type {IHTML5Properties} from "../../../lib/types/html5";
     import type {DESIGN_POINTS_ARGUMENT} from "../../../lib/types/points";
 
     import {map_data_attributes, map_global_attributes} from "../../../lib/util/attributes";
 
-    export let element: HTMLElement | null = null;
+    type $$Props = {
+        element?: HTMLDivElement;
 
-    let _class: string = "";
-    export let id: string = "";
-    export let name: string = "";
-    export let style: string = "";
-    export let tabindex: number | string = "";
-    export let title: string = "";
+        span?: DESIGN_POINTS_ARGUMENT;
+        span_x?: DESIGN_POINTS_ARGUMENT;
+        span_y?: DESIGN_POINTS_ARGUMENT;
+    } & IHTML5Properties &
+        IGlobalProperties;
 
+    export let element: $$Props["element"] = undefined;
+
+    let _class: $$Props["class"] = "";
     export {_class as class};
 
-    export let hidden: DESIGN_HIDDEN_ARGUMENT = false;
-
-    export let span: DESIGN_POINTS_ARGUMENT | undefined = undefined;
-    export let span_x: DESIGN_POINTS_ARGUMENT | undefined = undefined;
-    export let span_y: DESIGN_POINTS_ARGUMENT | undefined = undefined;
+    export let span: $$Props["span"] = undefined;
+    export let span_x: $$Props["span_x"] = undefined;
+    export let span_y: $$Props["span_y"] = undefined;
 </script>
 
 <div