+++ src/components/utilities/contextbackdrop/ContextBackdrop.svelte
@@ -3,70 +3,46 @@
      * TODO: Stories (?)
      */
 
+    import type {IGlobalProperties} from "../../../lib/types/global";
+    import type {IHTML5Properties} from "../../../lib/types/html5";
+    import type {IIntrinsicProperties} from "../../../lib/types/sizings";
+
     import {get_id_context} from "../../../lib/stores/id";
     import {get_state_context} from "../../../lib/stores/state";
     import {scrolllock} from "../../../lib/stores/scrolllock";
 
-    import type {DESIGN_HIDDEN_ARGUMENT} from "../../../lib/types/hidden";
-    import type {DESIGN_INTRINSIC_SIZING_ARGUMENT} from "../../../lib/types/sizings";
-    import type {DESIGN_SPACING_ARGUMENT} from "../../../lib/types/spacings";
-
     import {map_global_attributes} from "../../../lib/util/attributes";
     import {IS_BROWSER} from "../../../lib/util/environment";
 
-    export let element: HTMLElement | null = null;
+    type $$Props = {
+        element?: HTMLDivElement | HTMLLabelElement;
 
-    let _class: string = "";
-    export let id: string = "";
-    export let name: string = "";
-    export let style: string = "";
-    export let tabindex: number | string = "";
-    export let title: string = "";
+        dismissible?: boolean;
+    } & IHTML5Properties &
+        IGlobalProperties &
+        IIntrinsicProperties;
 
+    export let element: $$Props["element"] = undefined;
+
+    let _class: $$Props["class"] = "";
     export {_class as class};
 
-    export let hidden: DESIGN_HIDDEN_ARGUMENT = false;
-
-    export let margin: DESIGN_SPACING_ARGUMENT | undefined = undefined;
-
-    export let margin_x: DESIGN_SPACING_ARGUMENT | undefined = undefined;
-    export let margin_y: DESIGN_SPACING_ARGUMENT | undefined = undefined;
-
-    export let margin_top: DESIGN_SPACING_ARGUMENT | undefined = undefined;
-    export let margin_left: DESIGN_SPACING_ARGUMENT | undefined = undefined;
-    export let margin_bottom: DESIGN_SPACING_ARGUMENT | undefined = undefined;
-    export let margin_right: DESIGN_SPACING_ARGUMENT | undefined = undefined;
-
-    export let padding: DESIGN_SPACING_ARGUMENT | undefined = undefined;
-
-    export let padding_x: DESIGN_SPACING_ARGUMENT | undefined = undefined;
-    export let padding_y: DESIGN_SPACING_ARGUMENT | undefined = undefined;
-
-    export let padding_top: DESIGN_SPACING_ARGUMENT | undefined = undefined;
-    export let padding_left: DESIGN_SPACING_ARGUMENT | undefined = undefined;
-    export let padding_bottom: DESIGN_SPACING_ARGUMENT | undefined = undefined;
-    export let padding_right: DESIGN_SPACING_ARGUMENT | undefined = undefined;
-
-    export let height: DESIGN_INTRINSIC_SIZING_ARGUMENT | undefined = undefined;
-    export let max_height: DESIGN_INTRINSIC_SIZING_ARGUMENT | undefined = undefined;
-    export let min_height: DESIGN_INTRINSIC_SIZING_ARGUMENT | undefined = undefined;
-
-    export let width: DESIGN_INTRINSIC_SIZING_ARGUMENT | undefined = undefined;
-    export let max_width: DESIGN_INTRINSIC_SIZING_ARGUMENT | undefined = undefined;
-    export let min_width: DESIGN_INTRINSIC_SIZING_ARGUMENT | undefined = undefined;
-
-    export let dismissible: boolean = false;
+    export let dismissible: $$Props["dismissible"] = false;
 
     const _logic_id = get_id_context();
     const _scrolllock = scrolllock();
     const _state = get_state_context();
 
-    $: {
-        if (dismissible && !_logic_id) {
-            throw new Error(
-                "bad property 'dismissable' to 'ContextBackdrop' (Svelte Store was not set as a context)"
-            );
-        }
+    if (!_state) {
+        throw new ReferenceError(
+            "bad initialization to `ContextBackdrop` (failed to get `state` Svelte Store from context)"
+        );
+    }
+
+    if (dismissible && !_logic_id) {
+        throw new ReferenceError(
+            "bad initialization to `ContextBackdrop` (failed to get `id` Svelte Store from context)"
+        );
     }
 
     $: if (IS_BROWSER) $_scrolllock = $_state;