+++ src/lib/types/html5.ts
@@ -1,0 +1,8 @@
+export interface IHTML5Properties {
+    class?: string;
+    id?: string;
+    name?: string;
+    style?: string;
+    tabindex?: number | string;
+    title?: string;
+}